+++ src/main/java/com/steadybit/mcp/SteadybitMcpServerApplication.java
@@ -5,7 +5,6 @@
 package com.steadybit.mcp;
 
 import com.steadybit.mcp.tools.ActionTools;
-import com.steadybit.mcp.tools.ApiTools;
 import com.steadybit.mcp.tools.ExecutionTools;
 import com.steadybit.mcp.tools.ExperimentTools;
 import org.springframework.ai.tool.ToolCallbackProvider;
@@ -53,12 +52,7 @@ public class SteadybitMcpServerApplication {
     }
 
     @Bean
-    public ApiTools apiTools(RestClient restClient) {
-        return new ApiTools(restClient);
-    }
-
-    @Bean
-    public ToolCallbackProvider weatherTools(ExperimentTools experimentTools, ExecutionTools executionTools, ActionTools actionTools, ApiTools apiTools) {
-        return MethodToolCallbackProvider.builder().toolObjects(experimentTools, executionTools, actionTools, apiTools).build();
+    public ToolCallbackProvider weatherTools(ExperimentTools experimentTools, ExecutionTools executionTools, ActionTools actionTools) {
+        return MethodToolCallbackProvider.builder().toolObjects(experimentTools, executionTools, actionTools).build();
     }
 }