+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+- Removed `get_api_spec` tool, content gets truncated, and it's not really MCP agnostic / helpful.
+
 ## v0.0.6
 
 - Added `get_api_spec` tool to retrieve the API specification of the Steadybit platform